net/rsync: Fix distinfo file generation
Fix a long standing issue in distinfo file generation caused
by the way patch distfiles are handled.

In the previous code, every time a patch was activated, it adds its own
entry into the DISTFILES variable, but because all the patches comes
from the same distfile, we ended with duplicate entries in the list.
As a consequence, the make makesum target generates a distinfo file with
duplicate entries, causing issues to the builder who tries to fetch and untar
the same file in paralel.

Now we add the distfile to the list if one of the options is activated,
avoiding duplication in the DISTFILES list.

This fix has no impact on the generated package, so no PORTREVISION
bump is required.